# backend/openapi.yaml
openapi: 3.0.3
info:
  title: QuantDog API
  version: 1.0.0
  description: QuantDog stock market analysis API with technical indicators, sentiment analysis, and AI-powered research capabilities.
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.quantdog.com
    description: Production server
paths:
...
  /api/v1/instruments/{symbol}/bars:
    get:
      summary: Get OHLCV bars
      description: Retrieve daily OHLCV bars for a symbol
      parameters:
        - in: path
          name: symbol
          schema:
            type: string
          required: true
          description: Stock symbol
        - in: query
          name: start
          schema:
            type: string
            format: date
          required: false
          description: Start date (YYYY-MM-DD)
        - in: query
          name: end
          schema:
            type: string
            format: date
          required: false
          description: End date (YYYY-MM-DD)
        - in: query
          name: adjusted
          schema:
            type: boolean
          required: false
          description: Use adjusted prices
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 5000
          required: false
          description: Maximum rows to return
      responses:
        '200':
          description: Bars retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BarsResponse'
  /api/v1/instruments/{symbol}/indicators:
    get:
      summary: Get technical indicators
      description: Calculate and retrieve technical indicators for a symbol
      parameters:
        - in: path
          name: symbol
          schema:
            type: string
          required: true
          description: Stock symbol
        - in: query
          name: start
          schema:
            type: string
            format: date
          required: false
          description: Start date (YYYY-MM-DD)
        - in: query
          name: end
          schema:
            type: string
            format: date
          required: false
          description: End date (YYYY-MM-DD)
        - in: query
          name: adjusted
          schema:
            type: boolean
          required: false
          description: Use adjusted prices
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 252
          required: false
          description: Maximum bars for calculation
      responses:
        '200':
          description: Indicators calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndicatorsResponse'
        '404':
          description: No bars found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
...
  /api/v1/market/technical:
    post:
      summary: Technical analysis
      description: Perform comprehensive technical analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketAnalysisRequest'
      responses:
        '200':
          description: Analysis completed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/market/intel:
    post:
      summary: Market intelligence
      description: Get news and Twitter sentiment analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketAnalysisRequest'
      responses:
        '200':
          description: Intelligence gathered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketIntelResponse'
  /api/v1/market/macro:
    post:
      summary: Macro analysis
      description: Get macro environment analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketAnalysisRequest'
      responses:
        '200':
          description: Macro analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MacroAnalysisResponse'
...
  /api/v1/research/runs:
    post:
      summary: Create research run
      description: Create a new AI-powered research run (requires RESEARCH_ENABLED=true)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchRunRequest'
      responses:
        '202':
          description: Research run created and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchRunResponse'
        '404':
          description: Research feature disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
  /api/v1/research/runs/{run_id}:
    get:
      summary: Get research run status
      description: Get research run status and progress
      parameters:
        - in: path
          name: run_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Run status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchRunStatusResponse'
  /api/v1/research/runs/{run_id}/result:
    get:
      summary: Get research run result
      description: Get final research run results including all agent outputs
      parameters:
        - in: path
          name: run_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Run results retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchRunResultResponse'
components:
  schemas:
    EnvelopeError:
      type: object
      required:
        - code
        - msg
        - error
      properties:
        code:
          type: integer
          enum: [0]
        msg:
          type: string
        error:
          type: object
          required:
            - type
            - detail
          properties:
            type:
              type: string
            detail:
              type: string
...
    EnvelopeSuccess:
      type: object
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          enum: [1]
        msg:
          type: string
        data:
          type: object
    BarsResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                bars:
                  type: array
                  items:
                    $ref: '#/components/schemas/Bar'
                count:
                  type: integer
    Bar:
      type: object
      properties:
        symbol:
          type: string
        bar_date:
          type: string
          format: date
        ts_utc:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: integer
        adjusted:
          type: boolean
        source:
          type: string
    Indicators:
      type: object
      properties:
        sma20:
          type: number
          nullable: true
        sma50:
          type: number
          nullable: true
        rsi14:
          type: number
          nullable: true
        macd:
          type: number
          nullable: true
        macd_signal:
          type: number
          nullable: true
        macd_histogram:
          type: number
          nullable: true
        recent_high:
          type: number
          nullable: true
        recent_low:
          type: number
          nullable: true
        last_close:
          type: number
          nullable: true
    IndicatorsResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                symbol:
                  type: string
                bars_count:
                  type: integer
                indicators:
                  $ref: '#/components/schemas/Indicators'
...
    MarketAnalysisRequest:
      type: object
      required:
        - symbol
      properties:
        symbol:
          type: string
        limit:
          type: integer
          default: 20
    MarketIntelResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                symbol:
                  type: string
                news_count:
                  type: integer
                twitter_count:
                  type: integer
                sentiment:
                  type: string
                sentiment_score:
                  type: number
                source_status:
                  type: object
    MacroAnalysisResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                symbol:
                  type: string
                macro_theme:
                  type: string
                  enum: [growth, rates, inflation, liquidity, none]
                yield_10y:
                  type: number
                  nullable: true
                fed_rate:
                  type: number
                  nullable: true
                cpi:
                  type: number
                  nullable: true
                market_sentiment:
                  type: string
                  nullable: true
...
    ResearchRunRequest:
      type: object
      required:
        - symbol
      properties:
        symbol:
          type: string
        horizon:
          type: string
          enum: [1d, 1w, 1m, 3m, 6m, 1y]
          default: 1w
    ResearchRunResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                run_id:
                  type: string
                  format: uuid
                symbol:
                  type: string
                horizon:
                  type: string
                status:
                  type: string
                  enum: [pending, running, completed, failed]
    ResearchRunStatusResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                run_id:
                  type: string
                symbol:
                  type: string
                status:
                  type: string
                requested_at:
                  type: string
                  format: date-time
                started_at:
                  type: string
                  format: date-time
                completed_at:
                  type: string
                  format: date-time
                  nullable: true
                progress:
                  type: object
    ResearchRunResultResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                run_id:
                  type: string
                symbol:
                  type: string
                status:
                  type: string
                final_decision:
                  type: string
                  enum: [BUY, SELL, HOLD]
                  nullable: true
                final_confidence:
                  type: integer
                  nullable: true
                agent_outputs:
                  type: array
                  items:
                    type: object